DDF-3741 Pull out sort layout view
