+++ ui/packages/catalog-ui-search/src/main/webapp/component/component.less
@@ -122,6 +122,7 @@
 @import 'dropdown/result-sort/dropdown.result-sort.less';
 @import 'result-sort/result-sort';
 @import 'sort-item/sort-item';
+@import 'sort/sort';
 @import 'query-schedule/query-schedule';
 @import 'radio/radio';
 @import 'alert-settings/alert-settings';
