+++ ui/packages/catalog-ui-search/src/main/webapp/component/sort-item/sort-item.less
@@ -30,10 +30,6 @@
     margin-bottom: @minimumSpacing;
   }
 
-  > .sort-add {
-    width: 100%;
-  }
-
 }
 
 @{customElementNamespace}sort-item + @{customElementNamespace}sort-item {
