+++ ui/packages/catalog-ui-search/src/main/webapp/component/sort/sort.less
@@ -1,0 +1,9 @@
+@{customElementNamespace}sort {
+  display: block;
+  width: 100%;
+  overflow: hidden;
+
+  .sort-add {
+    width: 100%;
+  }
+}
